The ratio of vehicles is increasing day by day. This leads to many problems like traffic congestion, insufficient parking space, etc. As the demand of parking area and vehicles is constantly increasing, it consumes a lot of time and fuel to find an empty parking slot.
The ratio of vehicles is increasing day by day. This leads to many problems like traffic congestion, insufficient parking space, etc. As the demand of parking area and vehicles is constantly increasing, it consumes a lot of time and fuel to find an empty parking slot. To overcome this problem, the proposed system (Automatic Parking System) can provide driver, the information whether the slot is empty or not. When a car will arrive, it will automatically open the gate only if there is an empty parking slot available. This will reduce the efforts of the driver to find the empty slot in the parking area.

Co-Authored-By: jsjeelshah <jsjeelshah@users.noreply.github.com>
Co-Authored-By: kinjal-111 <kinjal-111@users.noreply.github.com>